update
was getting a seg error with uint16_t.  reverted back to uint32_t and apply a 0xFFFF mask after the counter

+++ include/rogue/protocols/packetizer/Controller.h
@@ -46,7 +46,7 @@ namespace rogue {
                uint32_t appIndex_;
                uint32_t tranIndex_;
                bool     transSof_[256];
-               uint16_t tranCount_[256];
+               uint32_t tranCount_[256];
                uint32_t crcInit_[256];
                uint8_t  tranDest_;
                uint32_t dropCount_;
